+++ .eslintrc.js
@@ -111,7 +111,8 @@ module.exports = {
                 "no-unused-vars": 0,
                 "@typescript-eslint/no-unused-vars": 2,
                 "semi": 0,
-                "@typescript-eslint/semi": 2
+                "@typescript-eslint/semi": 2,
+                "@typescript-eslint/no-empty-function": 0
             }
         }
     ]
